+++ spring-messaging/src/main/java/org/springframework/messaging/rsocket/annotation/support/ClientResponderFactory.java
@@ -22,6 +22,7 @@ import io.rsocket.RSocketFactory;
 
 import org.springframework.messaging.handler.invocation.reactive.ArgumentResolverConfigurer;
 import org.springframework.messaging.handler.invocation.reactive.ReturnValueHandlerConfigurer;
+import org.springframework.messaging.rsocket.MetadataExtractor;
 import org.springframework.messaging.rsocket.RSocketStrategies;
 import org.springframework.util.RouteMatcher;
