+++ spring-messaging/src/test/java/org/springframework/messaging/rsocket/DefaultMetadataExtractorTests.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.messaging.rsocket.annotation.support;
+package org.springframework.messaging.rsocket;
 
 import java.time.Duration;
 import java.util.Map;
@@ -31,16 +31,13 @@ import reactor.core.publisher.Mono;
 import org.springframework.core.codec.CharSequenceEncoder;
 import org.springframework.core.codec.StringDecoder;
 import org.springframework.core.io.buffer.DataBufferFactory;
-import org.springframework.messaging.rsocket.LeakAwareNettyDataBufferFactory;
-import org.springframework.messaging.rsocket.RSocketRequester;
-import org.springframework.messaging.rsocket.RSocketStrategies;
 import org.springframework.util.Assert;
 import org.springframework.util.MimeType;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.springframework.messaging.rsocket.annotation.support.MessagingRSocket.COMPOSITE_METADATA;
-import static org.springframework.messaging.rsocket.annotation.support.MessagingRSocket.ROUTING;
-import static org.springframework.messaging.rsocket.annotation.support.MetadataExtractor.ROUTE_KEY;
+import static org.springframework.messaging.rsocket.MetadataExtractor.COMPOSITE_METADATA;
+import static org.springframework.messaging.rsocket.MetadataExtractor.ROUTE_KEY;
+import static org.springframework.messaging.rsocket.MetadataExtractor.ROUTING;
 import static org.springframework.util.MimeTypeUtils.TEXT_HTML;
 import static org.springframework.util.MimeTypeUtils.TEXT_PLAIN;
 import static org.springframework.util.MimeTypeUtils.TEXT_XML;
